Rewrite FlightSQL client in terms of the mid level client
