+++ arrow-flight/tests/client.rs
@@ -18,6 +18,7 @@
 //! Integration test for "mid level" Client
 
 mod common {
+    // Use common mock server
     pub mod server;
 }
 use arrow_array::{RecordBatch, UInt64Array};
@@ -26,17 +27,17 @@ use arrow_flight::{
     HandshakeResponse, Ticket,
 };
 use bytes::Bytes;
-use common::server::TestFlightServer;
-use futures::{Future, TryStreamExt};
-use tokio::{net::TcpListener, task::JoinHandle};
+use common::server::{
+    do_test,
+    expect_status,
+    TestFlightServer
+};
+use futures::{TryStreamExt};
 use tonic::{
-    transport::{Channel, Uri},
     Status,
 };
 
-use std::{net::SocketAddr, sync::Arc, time::Duration};
-
-const DEFAULT_TIMEOUT_SECONDS: u64 = 30;
+use std::sync::Arc;
 
 #[tokio::test]
 async fn test_handshake() {
@@ -257,137 +258,4 @@ async fn test_do_get_error_in_record_batch_stream() {
         assert_eq!(test_server.take_do_get_request(), Some(ticket));
     })
     .await;
-}
-
-/// Runs the future returned by the function,  passing it a test server and client
-async fn do_test<F, Fut>(f: F)
-where
-    F: Fn(TestFlightServer, FlightClient) -> Fut,
-    Fut: Future<Output = ()>,
-{
-    let test_server = TestFlightServer::new();
-    let fixture = TestFixture::new(&test_server).await;
-    let client = FlightClient::new(fixture.channel().await);
-
-    // run the test function
-    f(test_server, client).await;
-
-    // cleanly shutdown the test fixture
-    fixture.shutdown_and_wait().await
-}
-
-fn expect_status(error: FlightError, expected: Status) {
-    let status = if let FlightError::Tonic(status) = error {
-        status
-    } else {
-        panic!("Expected FlightError::Tonic, got: {:?}", error);
-    };
-
-    assert_eq!(
-        status.code(),
-        expected.code(),
-        "Got {:?} want {:?}",
-        status,
-        expected
-    );
-    assert_eq!(
-        status.message(),
-        expected.message(),
-        "Got {:?} want {:?}",
-        status,
-        expected
-    );
-    assert_eq!(
-        status.details(),
-        expected.details(),
-        "Got {:?} want {:?}",
-        status,
-        expected
-    );
-}
-
-/// Creates and manages a running TestServer with a background task
-struct TestFixture {
-    /// channel to send shutdown command
-    shutdown: Option<tokio::sync::oneshot::Sender<()>>,
-
-    /// Address the server is listening on
-    addr: SocketAddr,
-
-    // handle for the server task
-    handle: Option<JoinHandle<Result<(), tonic::transport::Error>>>,
-}
-
-impl TestFixture {
-    /// create a new test fixture from the server
-    pub async fn new(test_server: &TestFlightServer) -> Self {
-        // let OS choose a a free port
-        let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-
-        println!("Listening on {addr}");
-
-        // prepare the shutdown channel
-        let (tx, rx) = tokio::sync::oneshot::channel();
-
-        let server_timeout = Duration::from_secs(DEFAULT_TIMEOUT_SECONDS);
-
-        let shutdown_future = async move {
-            rx.await.ok();
-        };
-
-        let serve_future = tonic::transport::Server::builder()
-            .timeout(server_timeout)
-            .add_service(test_server.service())
-            .serve_with_incoming_shutdown(
-                tokio_stream::wrappers::TcpListenerStream::new(listener),
-                shutdown_future,
-            );
-
-        // Run the server in its own background task
-        let handle = tokio::task::spawn(serve_future);
-
-        Self {
-            shutdown: Some(tx),
-            addr,
-            handle: Some(handle),
-        }
-    }
-
-    /// Return a [`Channel`] connected to the TestServer
-    pub async fn channel(&self) -> Channel {
-        let url = format!("http://{}", self.addr);
-        let uri: Uri = url.parse().expect("Valid URI");
-        Channel::builder(uri)
-            .timeout(Duration::from_secs(DEFAULT_TIMEOUT_SECONDS))
-            .connect()
-            .await
-            .expect("error connecting to server")
-    }
-
-    /// Stops the test server and waits for the server to shutdown
-    pub async fn shutdown_and_wait(mut self) {
-        if let Some(shutdown) = self.shutdown.take() {
-            shutdown.send(()).expect("server quit early");
-        }
-        if let Some(handle) = self.handle.take() {
-            println!("Waiting on server to finish");
-            handle
-                .await
-                .expect("task join error (panic?)")
-                .expect("Server Error found at shutdown");
-        }
-    }
-}
-
-impl Drop for TestFixture {
-    fn drop(&mut self) {
-        if let Some(shutdown) = self.shutdown.take() {
-            shutdown.send(()).ok();
-        }
-        if self.handle.is_some() {
-            // tests should properly clean up TestFixture
-            println!("TestFixture::Drop called prior to `shutdown_and_wait`");
-        }
-    }
 }
