+++ arrow-flight/tests/common/server.rs
@@ -15,19 +15,30 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use std::sync::{Arc, Mutex};
+use std::{
+    net::SocketAddr,
+    time::Duration,
+    sync::{Arc, Mutex},
+};
 
 use arrow_array::RecordBatch;
-use futures::{stream::BoxStream, TryStreamExt};
-use tonic::{metadata::MetadataMap, Request, Response, Status, Streaming};
+use futures::{stream::BoxStream, TryStreamExt, Future};
+use tonic::{
+    metadata::MetadataMap, Request, Response, Status, Streaming,
+    transport::{Channel, Uri},
+};
+use tokio::{net::TcpListener, task::JoinHandle};
 
 use arrow_flight::{
     encode::FlightDataEncoderBuilder,
     error::FlightError,
     flight_service_server::{FlightService, FlightServiceServer},
     Action, ActionType, Criteria, Empty, FlightData, FlightDescriptor, FlightInfo,
-    HandshakeRequest, HandshakeResponse, PutResult, SchemaResult, Ticket,
+    HandshakeRequest, HandshakeResponse, PutResult, SchemaResult, Ticket, FlightClient,
 };
+
+
+const DEFAULT_TIMEOUT_SECONDS: u64 = 30;
 
 #[derive(Debug, Clone)]
 /// Flight server for testing, with configurable responses
@@ -246,5 +257,139 @@ impl FlightService for TestFlightServer {
         _request: Request<Streaming<FlightData>>,
     ) -> Result<Response<Self::DoExchangeStream>, Status> {
         Err(Status::unimplemented("Implement do_exchange"))
+    }
+}
+
+
+/// Runs the future returned by the function,  passing it a test server and client
+pub async fn do_test<F, Fut>(f: F)
+where
+    F: Fn(TestFlightServer, FlightClient) -> Fut,
+    Fut: Future<Output = ()>,
+{
+    let test_server = TestFlightServer::new();
+    let fixture = TestFixture::new(&test_server).await;
+    let client = FlightClient::new(fixture.channel().await);
+
+    // run the test function
+    f(test_server, client).await;
+
+    // cleanly shutdown the test fixture
+    fixture.shutdown_and_wait().await
+}
+
+pub fn expect_status(error: FlightError, expected: Status) {
+    let status = if let FlightError::Tonic(status) = error {
+        status
+    } else {
+        panic!("Expected FlightError::Tonic, got: {:?}", error);
+    };
+
+    assert_eq!(
+        status.code(),
+        expected.code(),
+        "Got {:?} want {:?}",
+        status,
+        expected
+    );
+    assert_eq!(
+        status.message(),
+        expected.message(),
+        "Got {:?} want {:?}",
+        status,
+        expected
+    );
+    assert_eq!(
+        status.details(),
+        expected.details(),
+        "Got {:?} want {:?}",
+        status,
+        expected
+    );
+}
+
+/// Creates and manages a running TestServer with a background task
+pub struct TestFixture {
+    /// channel to send shutdown command
+    shutdown: Option<tokio::sync::oneshot::Sender<()>>,
+
+    /// Address the server is listening on
+    addr: SocketAddr,
+
+    // handle for the server task
+    handle: Option<JoinHandle<Result<(), tonic::transport::Error>>>,
+}
+
+impl TestFixture {
+    /// create a new test fixture from the server
+    pub async fn new(test_server: &TestFlightServer) -> Self {
+        // let OS choose a a free port
+        let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
+        let addr = listener.local_addr().unwrap();
+
+        println!("Listening on {addr}");
+
+        // prepare the shutdown channel
+        let (tx, rx) = tokio::sync::oneshot::channel();
+
+        let server_timeout = Duration::from_secs(DEFAULT_TIMEOUT_SECONDS);
+
+        let shutdown_future = async move {
+            rx.await.ok();
+        };
+
+        let serve_future = tonic::transport::Server::builder()
+            .timeout(server_timeout)
+            .add_service(test_server.service())
+            .serve_with_incoming_shutdown(
+                tokio_stream::wrappers::TcpListenerStream::new(listener),
+                shutdown_future,
+            );
+
+        // Run the server in its own background task
+        let handle = tokio::task::spawn(serve_future);
+
+        Self {
+            shutdown: Some(tx),
+            addr,
+            handle: Some(handle),
+        }
+    }
+
+    /// Return a [`Channel`] connected to the TestServer
+    pub async fn channel(&self) -> Channel {
+        let url = format!("http://{}", self.addr);
+        let uri: Uri = url.parse().expect("Valid URI");
+        Channel::builder(uri)
+            .timeout(Duration::from_secs(DEFAULT_TIMEOUT_SECONDS))
+            .connect()
+            .await
+            .expect("error connecting to server")
+    }
+
+    /// Stops the test server and waits for the server to shutdown
+    pub async fn shutdown_and_wait(mut self) {
+        if let Some(shutdown) = self.shutdown.take() {
+            shutdown.send(()).expect("server quit early");
+        }
+        if let Some(handle) = self.handle.take() {
+            println!("Waiting on server to finish");
+            handle
+                .await
+                .expect("task join error (panic?)")
+                .expect("Server Error found at shutdown");
+        }
+    }
+}
+
+impl Drop for TestFixture {
+    fn drop(&mut self) {
+        if let Some(shutdown) = self.shutdown.take() {
+            shutdown.send(()).ok();
+        }
+        if self.handle.is_some() {
+            // tests should properly clean up TestFixture
+            println!("TestFixture::Drop called prior to `shutdown_and_wait`");
+        }
     }
 }
